pyproject.toml
[project]
name = "python-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.128.0",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt[crypto]>=2.10.1",
    "pymongo>=4.16.0",
    "python-multipart>=0.0.21",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
    "ty>=0.0.11",
]

[tool.pytest.ini_options]
pythonpath = "src"
